style(linear_problem): drop redundant static_cast<size_t>(j) in flatten Pass 1
`ColIndex` has `strong::implicitly_convertible_to<Index>` (Index =
std::int32_t), and `std::vector::operator[]` takes `size_type` (size_t).
The implicit conversion chain ColIndex → Index → size_t is well-defined,
so the explicit `static_cast<size_t>` here added nothing semantically.

Removing it makes the index-into-matbeg expression read identically to
the Pass-2 counterpart at the same level of explicitness.  No behavioural
change — the same conversion chain runs at the bracket.

Investigated as a candidate fix for the pre-existing -O3 SEGFAULT in
the IEEE 9-bus line-losses tests (`test_kirchhoff_modes_ieee.cpp`), but
it is NOT the root cause: the underlying bug exists at `6cf57176` (pre-
15bb5c11), pre-dating today's work, and the cast removal does not
change the generated assembly's sign-extension behaviour.  Keeping the
removal as an independent stylistic cleanup.

+++ source/linear_problem.cpp
@@ -332,7 +332,7 @@ auto LinearProblem::flatten(const LpMatrixOptions& opts) -> FlatLinearProblem
     for (const auto& [j, v] : row.cmap) {
       assert(j >= 0 && "column index in row.cmap must be non-negative");
       if (eps < 0 || std::abs(v) > eps) [[likely]] {
-        ++matbeg[static_cast<size_t>(j)];
+        ++matbeg[j];
       }
     }
   }
